This PR was merged into the main branch.

Discussion
----------

[Platform] Use int for retry after

| Q             | A
| ------------- | ---
| Bug fix?      | not really
| New feature?  | no
| Docs?         | no <!-- required for new features -->
| Issues        | Fix #... <!-- prefix each issue number with "Fix #", no need to create an issue if none exist, explain below instead -->
| License       | MIT

According to RFC 7231 should be an integer
https://datatracker.ietf.org/doc/html/rfc7231#page-69

So let's avoid a useless float here.

Commits
-------

d9688121 Use int for retry after

+++ src/platform/src/Bridge/Anthropic/ResultConverter.php
@@ -42,7 +42,7 @@ class ResultConverter implements ResultConverterInterface
 
         if (429 === $response->getStatusCode()) {
             $retryAfter = $response->getHeaders(false)['retry-after'][0] ?? null;
-            $retryAfterValue = $retryAfter ? (float) $retryAfter : null;
+            $retryAfterValue = $retryAfter ? (int) $retryAfter : null;
             throw new RateLimitExceededException($retryAfterValue);
         }
